Responsive Home
Home screen was reshaped with responsive UI design.

+++ weather_app/lib/models/airQualityModel.dart
@@ -25,14 +25,14 @@ class AirQuality {
     final components = json['components'];
     return AirQuality(
       aqi: json['main']['aqi'], 
-      co: components['co'], 
-      no: components['no'], 
-      no2: components['no2'], 
-      o3: components['o3'], 
-      so2: components['so2'], 
-      pm2_5: components['pm2_5'], 
-      pm10: components['pm10'], 
-      nh3: components['nh3'],
+      co: (components['co']).toDouble(), 
+      no: (components['no']).toDouble(), 
+      no2: (components['no2']).toDouble(), 
+      o3: (components['o3']).toDouble(), 
+      so2: (components['so2']).toDouble(), 
+      pm2_5: (components['pm2_5']).toDouble(), 
+      pm10: (components['pm10']).toDouble(), 
+      nh3: (components['nh3']).toDouble(),
     );
   }
 }
